requirements.txt
accelerate==1.12.0
acres==0.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
astor==0.8.1
asttokens==3.0.1
attrs==25.4.0
bce-python-sdk==0.9.53
beautifulsoup4==4.14.2
bitsandbytes==0.48.2
blake3==1.0.8
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
compressed-tensors==0.12.2
configobj==5.0.9
configparser==7.2.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
depyf==0.20.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.1
fastapi==0.121.3
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.6.0
filelock==3.20.0
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
gguf==0.17.1
greenlet==3.2.4
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
imagesize==1.4.1
interegular==0.3.3
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langsmith==0.4.47
lark==1.2.2
lm-format-enforcer==0.11.3
loguru==0.7.3
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral_common==1.8.5
model-hosting-container-standards==0.1.9
modelscope==1.32.0
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==2.2.6
openai==2.8.1
openai-harmony==0.0.8
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.4
outlines_core==0.2.11
packaging==25.0
paddleocr==3.3.2
paddlepaddle==3.2.2
paddlex==3.3.10
pandas==2.3.3
parso==0.8.5
partial-json-parser==0.2.1.1.post7
pathlib==1.0.1
pdf2image==1.17.0
pillow==12.0.0
platformdirs==4.5.0
premailer==3.10.0
prettytable==3.17.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.2
prov==2.1.1
psutil==7.1.3
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycryptodome==3.23.0
pydantic==2.12.4
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.2.5
pypdfium2==5.0.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.4
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.4.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.29.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
sentry-sdk==2.45.0
setproctitle==1.3.7
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.50.0
supervisor==4.3.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1+cu128
torchaudio==2.9.1+cu128
torchvision==0.24.1+cu128
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.57.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.38.0
vllm==0.11.2
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
yarl==1.22.0
zstandard==0.25.0